% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{bexy}
\alias{bexy}
\title{Accurate Bayesian inference of sex chromosome karyotypes and sex-linked scaffolds from low-depth sequencing data}
\usage{
bexy(path = NULL, readMCMCTrace = FALSE)
}
\arguments{
\item{path}{The path where all the output files of BeXY are located. If NULL, the example files will be loaded.}

\item{readMCMCTrace}{If TRUE, read the full trace of the MCMC of all parameters (potentially slow).}
}
\value{
An object of type bexy.
}
\description{
Visualization of results produced by BeXY
}
