% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.addLegend.bexy}
\alias{.addLegend.bexy}
\title{Function to create a legend}
\usage{
.addLegend.bexy(x_coord, y_coord, colors)
}
\arguments{
\item{x_coord}{The x-coordinate for legend.}

\item{y_coord}{The y-coordinate for legend.}

\item{colors}{A vector of length three with the colors for aneuploid, XX and XY samples, respectively.}
}
\value{
No return value, called for side effects.
}
\description{
Function to create a legend
}
\keyword{internal}
