% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.addLegendAutosomalTrisomy.bexy}
\alias{.addLegendAutosomalTrisomy.bexy}
\title{Function to create a legend for autosomal trisomy}
\usage{
.addLegendAutosomalTrisomy.bexy(x_coord, y_coord, scaffoldName, colors)
}
\arguments{
\item{x_coord}{The x-coordinate for legend.}

\item{y_coord}{The y-coordinate for legend.}

\item{scaffoldName}{The name of the scaffold to plot (has to be an autosome).}

\item{colors}{A vector of length two with the colors for triploid and diploid samples, respectively.}
}
\value{
No return value, called for side effects.
}
\description{
Function to create a legend for autosomal trisomy
}
\keyword{internal}
