% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.getAllPossibleAutosomalTrisomy.bexy}
\alias{.getAllPossibleAutosomalTrisomy.bexy}
\title{Function that returns all possible autosomal trisomy states used in BeXY}
\usage{
.getAllPossibleAutosomalTrisomy.bexy(scaffoldName)
}
\arguments{
\item{scaffoldName}{A vector of size one with the name of a scaffold.}
}
\value{
A vector of characters.
}
\description{
Function that returns all possible autosomal trisomy states used in BeXY
}
\keyword{internal}
