% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.getAllPossibleScaffoldTypes.bexy}
\alias{.getAllPossibleScaffoldTypes.bexy}
\title{Function that returns all possible scaffold types used in BeXY}
\usage{
.getAllPossibleScaffoldTypes.bexy()
}
\value{
A vector of characters.
}
\description{
Function that returns all possible scaffold types used in BeXY
}
\keyword{internal}
