% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.getAllPossibleSexKaryotypes.bexy}
\alias{.getAllPossibleSexKaryotypes.bexy}
\title{Function that returns all possible sex karyotypes used in BeXY}
\usage{
.getAllPossibleSexKaryotypes.bexy()
}
\value{
A vector of characters.
}
\description{
Function that returns all possible sex karyotypes used in BeXY
}
\keyword{internal}
