% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.getColumns.bexy}
\alias{.getColumns.bexy}
\title{Function to extract specific columns from a data frame or matrix}
\usage{
.getColumns.bexy(what, name)
}
\arguments{
\item{what}{A matrix or data frame}

\item{name}{The name of the parameter to extract.}
}
\value{
A subset of the matrix or data frame containing only the columns whose names match the argument 'name'
}
\description{
Function to extract specific columns from a data frame or matrix
}
\keyword{internal}
