% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.getIndexScaffold.bexy}
\alias{.getIndexScaffold.bexy}
\title{Function that returns the index of a given scaffold}
\usage{
.getIndexScaffold.bexy(object, scaffoldName, takeIfNotFound)
}
\arguments{
\item{object}{A bexy object.}

\item{scaffoldName}{A string representing the name of a scaffold.}

\item{takeIfNotFound}{A string ("X-linked" or "Y-linked") that denotes the type of scaffold to use if scaffoldName could not be found}
}
\value{
A single integer representing the index of the scaffold.
}
\description{
Function that returns the index of a given scaffold
}
\keyword{internal}
