% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.getInputTernary.bexy}
\alias{.getInputTernary.bexy}
\title{Function that transforms state posteriors of sex karyotype to ternary-format}
\usage{
.getInputTernary.bexy(object)
}
\arguments{
\item{object}{A bexy object.}
}
\value{
A matrix with 3 rows: P(aneuploid), P(XX), P(XY).
}
\description{
Function that transforms state posteriors of sex karyotype to ternary-format
}
\keyword{internal}
