% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.getSamplesToHighlight.bexy}
\alias{.getSamplesToHighlight.bexy}
\title{Function to select samples that should be highlighted}
\usage{
.getSamplesToHighlight.bexy(object, sampleNamesToHighlight)
}
\arguments{
\item{object}{A bexy object.}

\item{sampleNamesToHighlight}{A vector of sample names that should be highlighted. If empty, samples are automatically highlighted based on the posterior probabilites.}
}
\value{
A vector of integers corresponding to indices of samples to highlight
}
\description{
Function to select samples that should be highlighted
}
\keyword{internal}
