% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.getSamplesToHighlightAutosomalTrisomy.bexy}
\alias{.getSamplesToHighlightAutosomalTrisomy.bexy}
\title{Function to select samples that should be highlighted for autosomal trisomy}
\usage{
.getSamplesToHighlightAutosomalTrisomy.bexy(object, ix, sampleNamesToHighlight)
}
\arguments{
\item{object}{A bexy object.}

\item{ix}{A list containing the index of a specific scaffold within z.}

\item{sampleNamesToHighlight}{A vector of sample names that should be highlighted. If empty, samples are automatically highlighted based on the posterior probabilites.}
}
\value{
A vector of integers corresponding to indices of samples to highlight
}
\description{
Function to select samples that should be highlighted for autosomal trisomy
}
\keyword{internal}
