% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.getScaffoldNames.bexy}
\alias{.getScaffoldNames.bexy}
\title{Function that returns the scaffold names of z}
\usage{
.getScaffoldNames.bexy(names_z, numSamples)
}
\arguments{
\item{names_z}{A character vector with the column names of z}

\item{numSamples}{The number of samples}
}
\value{
A vector of characters.
}
\description{
Function that returns the scaffold names of z
}
\keyword{internal}
