% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.getStringWidth.bexy}
\alias{.getStringWidth.bexy}
\title{Function to get the width of a string on the current plotting device in user coordinates}
\usage{
.getStringWidth.bexy(name)
}
\arguments{
\item{name}{A string.}
}
\value{
The width of the input string on the current plotting device in user coordinates.
}
\description{
Function to get the width of a string on the current plotting device in user coordinates
}
\keyword{internal}
