% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.openFile.bexy}
\alias{.openFile.bexy}
\title{Function to open a file and generate error message if it was not found}
\usage{
.openFile.bexy(path, files, pattern)
}
\arguments{
\item{path}{A file path.}

\item{files}{A vector of character strings corresponding to file names found in the path.}

\item{pattern}{A pattern to search for within 'files'.}
}
\value{
A file connection.
}
\description{
Function to open a file and generate error message if it was not found
}
\keyword{internal}
