% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.plotAutosomalTrisomy.bexy}
\alias{.plotAutosomalTrisomy.bexy}
\title{Function to plot counts of a specific autosome and color them according to BeXY posterior mode of autosomal trisomy}
\usage{
.plotAutosomalTrisomy.bexy(
  object,
  scaffoldName,
  plotCounts,
  colors,
  sampleNamesToHighlight,
  label,
  ...
)
}
\arguments{
\item{object}{A bexy object.}

\item{scaffoldName}{The name of the scaffold to plot (has to be an autosome)}

\item{plotCounts}{A boolean. If TRUE, the counts are plotted. If FALSE, the posterior probabilities are plotted.}

\item{colors}{A vector of length two with the colors for diploid and triploid samples, respectively.}

\item{sampleNamesToHighlight}{A vector of sample names that should be highlighted. If empty, samples are automatically highlighted based on the posterior probabilites.}

\item{label}{Boolean indicating whether samples should be labeled.}

\item{...}{Other parameters used for plotting}
}
\value{
No return value, called for side effects.
}
\description{
Function to plot counts of a specific autosome and color them according to BeXY posterior mode of autosomal trisomy
}
\keyword{internal}
