% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.plotTernaryAndTicTacToe.bexy}
\alias{.plotTernaryAndTicTacToe.bexy}
\title{Function to plot ternary plot with 7-cell plots displaying posterior probabilities of karyotypes}
\usage{
.plotTernaryAndTicTacToe.bexy(
  object,
  colors,
  sampleNamesToHighlight,
  label,
  addSquares,
  ...
)
}
\arguments{
\item{object}{A bexy object.}

\item{colors}{A vector of length three with the colors for aneuploid, XX and XY samples, respectively.}

\item{sampleNamesToHighlight}{A vector of sample names that should be highlighted. If empty, samples are automatically highlighted based on the posterior probabilities.}

\item{label}{Boolean indicating whether samples should be labeled.}

\item{addSquares}{Boolean indicating whether 7-cell posterior probability square plots should be drawn.}

\item{...}{Other parameters used for plotting}
}
\value{
No return value, called for side effects.
}
\description{
Function to plot ternary plot with 7-cell plots displaying posterior probabilities of karyotypes
}
\keyword{internal}
