% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.plotTicTacToeWithCoordinates.bexy}
\alias{.plotTicTacToeWithCoordinates.bexy}
\title{Function to plot a single 7-cell plot displaying posterior probabilities of karyotypes}
\usage{
.plotTicTacToeWithCoordinates.bexy(
  object,
  ix,
  label,
  x_coords,
  y_coords,
  last,
  colors
)
}
\arguments{
\item{object}{A bexy object.}

\item{ix}{The sample index.}

\item{label}{The sample name used as a label.}

\item{x_coords}{Left and right x-coordinates for 7-cell plot.}

\item{y_coords}{Bottom and top y-coordinates for 7-cell plot.}

\item{last}{Boolean indicating if this is the last sample to be plotted.}

\item{colors}{A vector of length three with the colors for aneuploid, XX and XY samples, respectively.}
}
\value{
No return value, called for side effects.
}
\description{
Function to plot a single 7-cell plot displaying posterior probabilities of karyotypes
}
\keyword{internal}
