% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.plotXVsY.bexy}
\alias{.plotXVsY.bexy}
\title{Function to plot X vs Y counts and color them according to BeXY posterior mode}
\usage{
.plotXVsY.bexy(object, colors, sampleNamesToHighlight, label, ...)
}
\arguments{
\item{object}{A bexy object.}

\item{colors}{A vector of length three with the colors for aneuploid, XX and XY samples, respectively.}

\item{sampleNamesToHighlight}{A vector of sample names that should be highlighted. If empty, samples are automatically highlighted based on the posterior probabilites.}

\item{label}{Boolean indicating whether samples should be labeled.}

\item{...}{Other parameters used for plotting}
}
\value{
No return value, called for side effects.
}
\description{
Function to plot X vs Y counts and color them according to BeXY posterior mode
}
\keyword{internal}
