% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{.renameHeader.bexy}
\alias{.renameHeader.bexy}
\title{Function to reformat column names}
\usage{
.renameHeader.bexy(file, pattern)
}
\arguments{
\item{file}{A file whose header should be renamed.}

\item{pattern}{Where to split.}
}
\value{
A new file with re-formatted column names.
}
\description{
Function to reformat column names
}
\keyword{internal}
