% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{getPosteriorMeanRho}
\alias{getPosteriorMeanRho}
\title{Getting posterior mean rho (ploidy ratio parameter) for each scaffold}
\usage{
getPosteriorMeanRho(object)
}
\arguments{
\item{object}{A bexy object.}
}
\value{
A numeric vector containing the posterior mean of rho for each scaffold
}
\description{
Getting posterior mean rho (ploidy ratio parameter) for each scaffold
}
\examples{
bex <- bexy()
getPosteriorMeanRho(bex)
}
\seealso{
\code{\link{bexy}}
}
