% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{getPosteriorModeScaffoldTypes}
\alias{getPosteriorModeScaffoldTypes}
\title{Getting the posterior mode for each scaffold type}
\usage{
getPosteriorModeScaffoldTypes(object, threshold_certainty = 0.9)
}
\arguments{
\item{object}{A bexy object.}

\item{threshold_certainty}{The threshold for certainty on the posterior probabilities. All scaffolds that have a posterior probability less than this threshold are classified as 'uncertain'.}
}
\value{
A character vector containing the scaffold type classification for each scaffold
}
\description{
Getting the posterior mode for each scaffold type
}
\examples{
bex <- bexy()
getPosteriorModeScaffoldTypes(bex)
}
\seealso{
\code{\link{bexy}}
}
