% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{getPosteriorModeSexKaryotypes}
\alias{getPosteriorModeSexKaryotypes}
\title{Getting the posterior mode for each sex karyotype}
\usage{
getPosteriorModeSexKaryotypes(object, threshold_certainty = 0.9)
}
\arguments{
\item{object}{A bexy object.}

\item{threshold_certainty}{The threshold for certainty on the posterior probabilities. All samples that have a posterior probability less than this threshold are classified as 'uncertain'.}
}
\value{
A character vector containing the sex karyotype classification for each sample
}
\description{
Getting the posterior mode for each sex karyotype
}
\examples{
bex <- bexy()
getPosteriorModeSexKaryotypes(bex)
}
\seealso{
\code{\link{bexy}}
}
