% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{print.bexy}
\alias{print.bexy}
\title{Printing a bexy object}
\usage{
\method{print}{bexy}(x, ...)
}
\arguments{
\item{x}{A bexy object.}

\item{...}{Additional parameters passed to print functions.}
}
\value{
No return value, called for side effects.
}
\description{
Printing a bexy object
}
\examples{
bex <- bexy()
print(bex)
}
\seealso{
\code{\link{bexy}}
}
