% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexy.R
\name{summary.bexy}
\alias{summary.bexy}
\title{Summarizing a bexy object}
\usage{
\method{summary}{bexy}(object, ...)
}
\arguments{
\item{object}{A bexy object.}

\item{...}{Additional parameters passed to summary functions.}
}
\value{
No return value, called for side effects.
}
\description{
Summarizing a bexy object
}
\examples{
bex <- bexy()
print(bex)
}
\seealso{
\code{\link{bexy}}
}
